requirements.txt
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
mysqlclient==1.4.4

